Remove the useless (and ugly) tools/ folder
