--- lib/bump/cli/tools/definition.rb
@@ -1,44 +1,0 @@
-require 'bump/cli/tools/parser'
-require 'bump/cli/tools/references'
-
-module Bump
-  class CLI
-    module Tools
-      class Definition
-        def initialize(path, import_external_references: false)
-          @path = path
-          @import_external_references = import_external_references
-        end
-
-        def prepare
-          if !import_external_references
-            read_file
-          else
-            parse_file_and_import_external_references
-          end
-        end
-
-        private
-
-        attr_reader :path, :import_external_references
-
-        def read_file
-          open(path).read
-        end
-
-        def parse_file_and_import_external_references
-          original_format, definition = parser.load(read_file)
-
-          references = References.new(definition, base_path: path)
-          references.import!
-
-          parser.dump(references.definition, original_format)
-        end
-
-        def parser
-          @parser ||= Parser.new
-        end
-      end
-    end
-  end
-end
