--- lib/bump/cli/tools/uuid.rb
@@ -1,13 +1,0 @@
-module Bump
-  class CLI
-    module Tools
-      class UUID
-        REGEXP = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/
-
-        def self.valid?(string)
-          string.to_s =~ REGEXP
-        end
-      end
-    end
-  end
-end
